+++ src/Utils.cpp
@@ -119,12 +119,12 @@ int rotateSide( int side, int orient )
   return side;
 }
 
-Vector3f sideToNormal( int side )
+Vector3d sideToNormal( int side )
 {
   int sideToRowRemap[ 6 ] = { 5, 2, 1, 4, 3, 0 };
 
   int index = cubeIndices[ sideToRowRemap[ side ] * 6 ];
-  return Vector3f( cubeNormals[ index ][ 0 ],
+  return Vector3d( cubeNormals[ index ][ 0 ],
                    cubeNormals[ index ][ 1 ],
                    cubeNormals[ index ][ 2 ] );
 }
